Separate files_fake
